backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1

# Resume parsing
pdfplumber==0.11.0
python-docx==1.1.2
PyPDF2==3.0.1

# LLM
google-generativeai==0.8.3

# RAG / Embeddings (sentence-transformers omitted — keyword fallback used instead,
# avoids pulling PyTorch ~800MB into the image)
faiss-cpu==1.8.0
numpy==1.26.4

# HTTP
httpx==0.27.0
